Add bind_dn_template to Active Directory instructions
The Active Directory authentication code path uses two accounts:
1) Read-only technical account to perform username -> userdn lookup
2) userdn + password to bind (to check the password is valid)

This code path is associated with the `lookup_dn=True` configuration
setting. When you have this setting in effect, you must also set
`bind_dn_template='{username}'`, otherwise the loop is not entered and
the code does not even attempt to bind at all. The code doesn't attempt
to perform "Step 2" and check if the password is valid.

Change the documentation to show the required configuration for
`bind_dn_template` when `lookup_dn=True` is in effect.