Update README instructions about configuration option bind_dn_template
The bind_dn_template configuration option has a poor default value. If
the default value (the empty list) is used, the ldapauthenticator code
will silently ignore the request to bind and not even attempt to
authenicate with the LDAP server at all. This is true regardless of how
the user has configured the `lookup_dn` option.